add explicit cast when returning vector size as uint32_t
Fixes vs8 warning messages

+++ include/bmx/as02/AS02Clip.h
@@ -71,7 +71,7 @@ public:
 public:
     mxfRational GetFrameRate() const { return mClipFrameRate; }
 
-    uint32_t GetNumTracks() const { return mTracks.size(); }
+    uint32_t GetNumTracks() const { return (uint32_t)mTracks.size(); }
     int64_t GetDuration();
 
     AS02Track* GetTrack(uint32_t track_index);
